Exclude IA OCR-test clones of UNC items
All of UNC's Scribed items are expected to be in the
`unclibraries` collection, and the search for UNC items
is now limited to items in the `unclibraries` collection.
IA cloned two of UNC's items for use in OCR-testing; these
items were only in the collection `tesseract_test`. They
still contained unc_bib_record_id values and such and
were causing it to appear as if there were problematic
duplicate items/metadata.